update to deploy and test code. Still getting bad acc unfortunately.

+++ proc_for_infer.py
@@ -7,7 +7,7 @@ def preproc(x):
     x = x - np.asarray(PREPROC_MINS)
     x = x / np.asarray(pre_incrs)
     x = x.round()
-    #x = x.astype("int8")
+    x = x.astype("int8")
     return x
 
 def postproc(x):
